Make Bloom filter stable across program runs
GetHashCode in .NET Core returns a different value for each app domain.
That's fine within the context of a single run, but it means we can't serialise
the filter and load it again later. Burned a lot of time on this..

Also added some methods for loading the filter from its consistuent parts
(may want to simplify the api, so kept it internal for now)